requirements.txt
arrow==1.2.3
asgiref==3.6.0
asttokens==2.2.1
atomicwrites==1.4.1
attrs==23.1.0
backcall==0.2.0
binaryornot==0.4.4
black==23.3.0
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cookiecutter==2.1.1
coverage==7.2.5
crispy-bootstrap5==0.7
cryptography==40.0.2
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.1
django-allauth==0.54.0
django-browser-reload==1.8.0
django-crispy-forms==2.0
django-debug-toolbar==4.1.0
django-extensions==3.2.1
django-tailwind==3.5.0
exceptiongroup==1.1.1
executing==1.2.0
factory-boy==3.2.1
Faker==18.9.0
flake8==6.0.0
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
ipython==8.13.2
ipython-genutils==0.2.0
jedi==0.18.2
Jinja2==3.1.2
jinja2-time==0.2.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mock==5.0.2
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.1
parso==0.8.3
pathspec==0.11.1
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.1
pluggy==1.0.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
pure-eval==0.2.2
py==1.11.0
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
Pygments==2.15.1
PyJWT==2.7.0
pyparsing==3.0.9
pytest==7.3.1
pytest-cov==4.0.0
pytest-django==4.5.2
pytest-factoryboy==2.5.1
python-dateutil==2.8.2
python-decouple==3.8
python-slugify==8.0.1
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
regex==2023.5.5
requests==2.30.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.0.2
wcwidth==0.2.6
